refactor: prayerCreation Dir 구조 변경 #25
- prayers 로 통일
