+++ src/domain/prayerUpdate/types/prayerUpdateItem.ts
@@ -1,3 +1,3 @@
-import type { PrayerCreationItem } from "../../prayerCreation/types/PrayerCreationItem";
+import type { PrayerCreationItem } from "../../prayers/types/PrayerCreationItem";
 
 export interface PrayerUpdateItem extends PrayerCreationItem {}
