+++ src/domain/prayers/types/params/createPrayerParams.ts
@@ -1,0 +1,6 @@
+import { PrayerCreationItem } from "./../PrayerCreationItem";
+
+export interface CreatePrayerParams {
+  title: string;
+  prayerList: PrayerCreationItem[];
+}
